#3 move localcopy to abstract
