+++ src/senzing/abstract/g2config_abstract.py
