+++ src/senzing/abstract/g2configmgr_abstract.py
