+++ src/senzing/abstract/g2diagnostic_abstract.py
