+++ src/senzing/abstract/g2engine_abstract.py
@@ -13,7 +13,7 @@ import json
 from abc import ABC, abstractmethod
 from typing import Any, Dict, Tuple, Union, cast
 
-from senzing.localcopy.g2engineflags import G2EngineFlags  # TODO: Fix this import mess
+from senzing.abstract.g2engineflags import G2EngineFlags  # TODO: Fix this import mess
 
 # Metadata
