+++ src/senzing/abstract/g2engineflags.py
