+++ src/senzing/abstract/g2hasher_abstract.py
