+++ src/senzing/abstract/g2product_abstract.py
