+++ src/senzing/g2diagnostic_grpc.py
@@ -11,8 +11,8 @@ from typing import Any, Dict, Type, Union
 
 import grpc
 
+from .abstract.g2diagnostic_abstract import G2DiagnosticAbstract
 from .g2helpers import new_exception
-from .localcopy.g2diagnostic_abstract import G2DiagnosticAbstract
 from .pb2_grpc import g2diagnostic_pb2, g2diagnostic_pb2_grpc
 
 # Metadata
