+++ src/senzing/g2product_grpc.py
@@ -11,8 +11,8 @@ from typing import Any, Dict, Type, Union
 
 import grpc
 
+from .abstract.g2product_abstract import G2ProductAbstract
 from .g2helpers import new_exception
-from .localcopy.g2product_abstract import G2ProductAbstract
 from .pb2_grpc import g2product_pb2, g2product_pb2_grpc
 
 # Metadata
